refactor(layout): layoutSize를 layout 하위로 이동
