+++ src/layout/Side/SideArea/SideArea.tsx
@@ -7,7 +7,7 @@ import useEventHandler from '../../../hooks/useEventHandler'
 import useLayoutState from '../../hooks/useLayoutState'
 import useResizingHandlers from '../../hooks/useResizingHandlers'
 import { document } from '../../../utils/domUtils'
-import { CONTENT_MIN_WIDTH, SIDE_MAX_WIDTH, SIDE_MIN_WIDTH } from '../../../constants/LayoutSizes'
+import { CONTENT_MIN_WIDTH, SIDE_MAX_WIDTH, SIDE_MIN_WIDTH } from '../../LayoutSizes'
 import useLayoutDispatch from '../../hooks/useLayoutDispatch'
 import LayoutActions from '../../redux/LayoutActions'
 import { Resizer, ScrollWrapper, SideAreaWrapper } from './SideArea.styled'
